// src/app/models/email/pendingcasesbysearch.ts
export interface IPendingCasesBySearch {
    workID: number;
    productivityID: number;
    dbMailAttachment: any[]; // Adjust the type as per your attachment model/interface
    attachmentMasterID: number;
    attachmentsPath: string | null;
    fileName: string | null;
    showExcelIcon: boolean;
    showPDFIcon: boolean;
    showJPEGIcon: boolean;
    showTXTIcon: boolean;
    showFileIcon: boolean;
    fileNumber: string | null;
    hblNumber: string | null;
    bookingNumber: string | null;
    emailClassID: number;
    emailCatID: number;
    emailCategory: string | null;
    receivedDate: string;
    creationDate: string;
    orderNumber: string | null;
    workStatus: string;
    workedBy: string;
    workStartDate: string;
    workFinishDate: string;
    firstActionID: number;
    lastActionID: number;
    expiryDate: string;
    emailID: number;
    inboxID: number;
    from: string | null;
    senderName: string | null;
    to: string | null;
    emailTo: string | null;
    emailFrom: string | null;
    cc: string | null;
    bCc: string | null;
    subject: string | null;
    dateTime: string;
    body: string | null;
    mainEmailID: number;
    empID: string | null;
    activityDate: string;
    activityStartTime: string;
    activityEndTime: string;
    idleTime: string | null;
    activityTotalTime: string | null;
    actionID: number | null;
    dueDate: string | null;
    reasonID: number;
    comments: string | null;
    emailSubCategory: string | null;
    categoryID: number;
    referenceNumber: string;
    incorrectTagging: boolean;
    statusID: number;
    accountNumber: string;
    mpan: string;
    futureDate: string;
    correctedSubCategoryID: number | null;
    isSubCategoryCorrect: boolean | null;
    graphId: number | null;
    categoryEmail: string | null;
    bulkRequest: boolean;
    numberOfBulkRequest: number;
    sentiment: string | null;
    spamOrJunk: string | null;
    gmtDateTime: string | null;
    contactTypeId: number;
    statusMessage: string | null;
  }
  